+++ src/Phan/CLI.php
@@ -47,7 +47,6 @@ use function getenv;
 use function in_array;
 use function is_array;
 use function is_executable;
-use function is_resource;
 use function is_string;
 use function min;
 use function phpversion;
@@ -83,7 +82,7 @@ class CLI
     /**
      * This should be updated to x.y.z-dev after every release, and x.y.z before a release.
      */
-    public const PHAN_VERSION = '5.0.0a3-dev';
+    public const PHAN_VERSION = '5.0.1-dev';
 
     /**
      * List of short flags passed to getopt
@@ -618,7 +617,7 @@ class CLI
                         throw new UsageException(\sprintf("Invalid arguments to --output: args=%s\n", StringUtil::jsonEncode($value)), EXIT_FAILURE);
                     }
                     $output_file = \fopen($value, 'w');
-                    if (!is_resource($output_file)) {
+                    if (!$output_file) {
                         throw new UsageException("Failed to open output file '$value'\n", EXIT_FAILURE, null, true);
                     }
                     $this->output = new StreamOutput($output_file);
@@ -849,7 +848,7 @@ class CLI
                         break;
                     }
                     $ast_version = (new ReflectionExtension('ast'))->getVersion();
-                    // In order to parse with AST version 80, 1.0.7+ is required
+                    // In order to parse with AST version 85, 1.0.11+ is required
                     if (\version_compare($ast_version, Config::MINIMUM_AST_EXTENSION_VERSION) < 0) {
                         Config::setValue('use_polyfill_parser', true);
                         break;
@@ -1532,13 +1531,13 @@ $init_help
  -b, --backward-compatibility-checks
   Check for potential PHP 5 -> PHP 7 BC issues
 
- --target-php-version {7.0,7.1,7.2,7.3,7.4,8.0,native}
+ --target-php-version {5.6,7.0,7.1,7.2,7.3,7.4,8.0,8.1,native}
   The PHP version that the codebase will be checked for compatibility against.
   For best results, the PHP binary used to run Phan should have the same PHP version.
   (Phan relies on Reflection for some param counts
    and checks for undefined classes/methods/functions)
 
- --minimum-target-php-version {7.0,7.1,7.2,7.3,7.4,8.0,native}
+ --minimum-target-php-version {5.6,7.0,7.1,7.2,7.3,7.4,8.0,8.1,native}
   The PHP version that will be used for feature/syntax compatibility warnings.
 
  -i, --ignore-undeclared
@@ -2396,7 +2395,7 @@ EOB
     {
         static $did_end = false;
         if ($did_end) {
-            // Overkill as a sanity check
+            // Overkill to prevent redundant output.
             return;
         }
         $did_end = true;
@@ -2405,7 +2404,7 @@ EOB
             return;
         }
         if (self::shouldShowProgress()) {
-            // Print a newline to stderr to visuall separate stderr from stdout
+            // Print a newline to stderr to visually separate stderr from stdout
             fwrite(STDERR, PHP_EOL);
             \fflush(\STDOUT);
         }
@@ -2689,7 +2688,7 @@ EOB
             \phan_output_ast_installation_instructions();
             exit(EXIT_FAILURE);
         }
-        self::sanityCheckAstVersion();
+        self::exitIfAstVersionIsInvalid();
 
         try {
             // Split up the opening PHP tag to fix highlighting in vim.
@@ -2734,7 +2733,7 @@ EOB
     /**
      * This duplicates the check in Bootstrap.php, in case opcache.file_cache has outdated information about whether extension_loaded('ast') is true. exists.
      */
-    private static function sanityCheckAstVersion(): void
+    private static function exitIfAstVersionIsInvalid(): void
     {
         $ast_version = (string)\phpversion('ast');
         if (\version_compare($ast_version, '1.0.0') <= 0) {
@@ -2745,7 +2744,8 @@ EOB
             // NOTE: We haven't loaded the autoloader yet, so these issue messages can't be colorized.
             \fprintf(
                 STDERR,
-                "ERROR: Phan 4.x requires php-ast 1.0.7+ because it depends on AST version 80. php-ast '%s' is installed." . PHP_EOL,
+                "ERROR: Phan 5.x requires php-ast %s+ because it depends on AST version 85. php-ast '%s' is installed." . PHP_EOL,
+                Config::MINIMUM_AST_EXTENSION_VERSION,
                 $ast_version
             );
             require_once __DIR__ . '/Bootstrap.php';